requirements.txt
langextract
google-generativeai
requests
beautifulsoup4
PyMuPDF
pandas
python-dotenv
boto3